packaging: build in debug mode instead of release mode
This must be done for the moment because of a bug in gtk4.
To reproduce this bug, build the project in release mode,
and run the app. Select your interface and wait for a few APs
to show up, and then, press the button "Last access point".
The cursor will jump to the first access point instead of
the last one. In debug mode, this problem is not happening.

Seen on Kali Linux 2025.2 arm64

Signed-off-by: Martin Olivier <martin.olivier@live.fr>